Case Study 2 - AI Agents/requirements.txt
python-dotenv
langchain
langchain-openai
langchain-mongodb
langchain-community
langchain-core
pydantic
psycopg2-binary # Für die PostgreSQL-Verbindung mit SQLDatabase